refactor jest test file to ts
